requirements.txt
# Web App Dependencies

Flask==1.1.0
requests==2.22.0

# Date Parsing

python-dateutil==2.8.0
pytz==2019.1


# Caching

redis==3.3.11
